Minimize main-thread work 최적화
